+++ lib/src/elements/xvi/xvi_file.dart
@@ -4,6 +4,7 @@ import 'package:collection/collection.dart';
 import 'package:mapiah/src/auxiliary/mp_numeric_aux.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/elements/mixins/mp_bounding_box_mixin.dart';
+import 'package:mapiah/src/elements/mixins/mp_thfile_reference_mixin.dart';
 import 'package:mapiah/src/elements/parts/th_length_unit_part.dart';
 import 'package:mapiah/src/elements/parts/th_position_part.dart';
 import 'package:mapiah/src/elements/parts/types/th_length_unit_type.dart';
@@ -12,7 +13,7 @@ import 'package:mapiah/src/elements/xvi/xvi_shot.dart';
 import 'package:mapiah/src/elements/xvi/xvi_sketchline.dart';
 import 'package:mapiah/src/elements/xvi/xvi_station.dart';
 
-class XVIFile with MPBoundingBoxMixin {
+class XVIFile with MPTHFileReferenceMixin, MPBoundingBoxMixin {
   String filename;
   double gridSizeLength;
   THLengthUnitPart gridSizeUnit;